Move cn to directory consistent with shadcn
