+++ developer-preview/src/lib/getPreviewData.test.tsx
@@ -1,4 +1,4 @@
-import { type ERC7730Schema } from "~/types";
+import { type ERC7730Schema } from "~/types/ERC7730Schema";
 import { getPreviewData } from "./getPreviewData";
 
 const exampleTransactionContext = {
